edition de pdf en synchrone

+++ composition-bouquets/backend/src/main/java/com/atelierchantdefleur/bouquetcomposer/model/rest/MateriauRest.java
@@ -7,6 +7,6 @@ public class MateriauRest {
 
     private Long id;
     private String nom;
-    private Integer prixUnitaire;
+    private float prixUnitaire;
     private FournisseurRest fournisseurRest;
 }
